Added example to enumerate node Ids level & contents

+++ src/test/java/net/sf/jsi/NullIndex.java
@@ -36,21 +36,21 @@ public class NullIndex implements SpatialIndex {
   }
 
   /**
-   * @see net.sf.jsi.SpatialIndex#nearest(Point, gnu.trove.TIntProcedure, float)
+   * @see net.sf.jsi.SpatialIndex#nearest(Point, gnu.trove.TIntProcedure, double)
    */
-  public void nearest(Point p, TIntProcedure v, float distance) {
+  public void nearest(Point p, TIntProcedure v, double distance) {
   }
 
   /**
-   * @see net.sf.jsi.SpatialIndex#nearestN(Point, gnu.trove.TIntProcedure, int, float)
+   * @see net.sf.jsi.SpatialIndex#nearestN(Point, gnu.trove.TIntProcedure, int, double)
    */
-  public void nearestN(Point p, TIntProcedure v, int n, float distance) {
+  public void nearestN(Point p, TIntProcedure v, int n, double distance) {
   }
  
   /**
-   * @see net.sf.jsi.SpatialIndex#nearestNUnsorted(Point, gnu.trove.TIntProcedure, int, float)
+   * @see net.sf.jsi.SpatialIndex#nearestNUnsorted(Point, gnu.trove.TIntProcedure, int, double)
    */
-  public void nearestNUnsorted(Point p, TIntProcedure v, int n, float distance) {
+  public void nearestNUnsorted(Point p, TIntProcedure v, int n, double distance) {
   }
   
   /**
